Reorganized some component code
