+++ packages/bvaughn-architecture-demo/components/sources/PointPanelTimeline.module.css
@@ -1,0 +1,95 @@
+.NextHitPointButton,
+.PreviousHitPointButton {
+  flex-grow: 0;
+  flex-shrink: 0;
+  background: none;
+  border: none;
+  margin: 0;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.PreviousHitPointButton,
+.NextHitPointButton {
+  height: 1.5rem;
+  width: 1.5rem;
+  border-radius: 1.5rem;
+  background-color: var(--background-color-tertiary-button);
+  color: var(--color-tertiary-button);
+  padding: 0;
+  transition: opacity 250ms;
+}
+
+.PreviousHitPointButton:disabled,
+.NextHitPointButton:disabled {
+  opacity: 0.5;
+  cursor: default;
+}
+
+.PreviousHitPointButtonIcon,
+.NextHitPointButtonIcon {
+  height: 1.5rem;
+  width: 1.5rem;
+}
+
+.HitPointsLabel {
+  background-color: var(--background-color-contrast-1);
+  font-size: var(--font-size-tiny);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 10ch;
+  height: 3ch;
+  border-radius: 1rem;
+}
+
+.HitPointTimeline,
+.HitPointTimelineDisabled {
+  position: relative;
+  background-color: var(--background-color-contrast-1);
+  flex: 1 1 auto;
+  border-radius: 0.5rem;
+  height: 0.5rem;
+  transition: opacity 250ms;
+}
+.HitPointTimelineDisabled {
+  opacity: 0.5;
+}
+
+.HitPointButton,
+.HitPointButtonCurrent {
+  position: absolute;
+  top: 0.25rem;
+  margin: 0;
+  margin-top: -0.75ch;
+  margin-left: -0.75ch;
+  padding: 0;
+  border: 0;
+  width: 1.5ch;
+  height: 1.5ch;
+  border-radius: 1.5ch;
+  outline: 0;
+}
+.HitPointButton:disabled,
+.HitPointButtonCurrent:disabled {
+  cursor: default;
+}
+.HitPointButton {
+  background-color: var(--color-brand);
+  z-index: 2;
+  cursor: pointer;
+}
+.HitPointButtonCurrent {
+  background-color: var(--color-current);
+  z-index: 3;
+}
+
+.CurrentTimeProgressBar {
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 0.5rem;
+  background-color: var(--color-brand);
+}
